first refactoring to add connection pool
